refactor(components): Guild Title v5
